cleaned up cron scripts, won't complain anymore
combsc

+++ scripts/stopCron.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
 
-kill -15 $(ps -e | grep "driver.exe" | grep -v grep | awk '{ print $1 }')
+if pgrep -x "driver.exe" > /dev/null
+then
+	kill -15 $(ps -e | grep "driver.exe" | grep -v grep | awk '{ print $1 }')
+fi
 crontab -r
